[#384] Updated creation of resume data on failure

+++ auto-setup/functions/autoSetupScriptInstall.sh
@@ -20,27 +20,35 @@ autoSetupScriptInstall() {
 
     # Execute scripts
     for script in ${scriptsToExecute}; do
-
-        # Unblock execution if set to true and script to be executed matches script to be retried
-        if [[ "${blockScriptExecution}" == "true" ]] && [[ ${retryScript} == ${script} ]]; then
-            blockScriptExecution="false"
-        else
-            log_debug "Skipping execution of script ${script} as already executed successfully prior to retry of component installation"
-
-        fi
-
-        # Execute script if there script execution is not blocked
-        if [[ "${blockScriptExecution}" != "true" ]]; then
-            log_info "Executing script \"${script}\" in directory ${installComponentDirectory}"
-            updateStorageClassIfNeeded "${installComponentDirectory}/${script}"
-            # Export retry data in case an error errors and the component installation needs to be retried
+   
+        if [[ ! -f ${installComponentDirectory}/${script} ]]; then
+            log_error "Install script ${installComponentDirectory}/${script} does not exist. Check your spelling in the \"metadata.json\" file and that it is checked in correctly into Git"
             autoSetupSaveRetryData "2" "main_scripts" "${script}" "${payload}"
-            . ${installComponentDirectory}/${script} || rc=$? && log_info "${installComponentDirectory}/${script} returned with rc=$rc"
-            if [[ ${rc} -ne 0 ]]; then
-                log_error "Execution of install script \"${script}\" ended in a non zero return code ($rc)"
-                exit 1
+            setRetryDataFailureState
+            exit 1
+        else
+
+            # Unblock execution if set to true and script to be executed matches script to be retried
+            if [[ "${blockScriptExecution}" == "true" ]] && [[ ${retryScript} == ${script} ]]; then
+                blockScriptExecution="false"
             else
-                autoSetupClearRetryData
+                log_debug "Skipping execution of script ${script} as already executed successfully prior to retry of component installation"
+
+            fi
+
+            # Execute script if there script execution is not blocked
+            if [[ "${blockScriptExecution}" != "true" ]]; then
+                log_info "Executing script \"${script}\" in directory ${installComponentDirectory}"
+                updateStorageClassIfNeeded "${installComponentDirectory}/${script}"
+                # Export retry data in case an error errors and the component installation needs to be retried
+                autoSetupSaveRetryData "2" "main_scripts" "${script}" "${payload}"
+                . ${installComponentDirectory}/${script} || rc=$? && log_info "${installComponentDirectory}/${script} returned with rc=$rc"
+                if [[ ${rc} -ne 0 ]]; then
+                    log_error "Execution of install script \"${script}\" ended in a non zero return code ($rc)"
+                    exit 1
+                else
+                    autoSetupClearRetryData
+                fi
             fi
         fi
     done
